\name{birk-package}
\alias{birk-package}
\alias{birk}
\docType{package}
\title{MA Birk functions}
\description{This is a compilation of functions that I found useful to make. It currently includes a unit of measurement conversion function and a few simple arithmetic functions.}
\details{
\tabular{ll}{
Package: \tab birk\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2014-11-07\cr
License: \tab GPL-2\cr
}}
\author{Matthew A. Birk \email{matthewabirk@gmail.com}}
\keyword{ package }