% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birk.R
\docType{package}
\encoding{UTF-8}
\name{birk}
\alias{birk}
\alias{birk-package}
\title{MA Birk's Functions}
\description{
Collection of tools to make R more convenient. Includes tools to summarize data using statistics not available with base R and manipulate objects for analyses.
}
\author{
Matthew A. Birk, \email{matthewabirk@gmail.com}
}

