% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\encoding{UTF-8}
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an R object which is coercible to one by as.vector(x, "numeric”).}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Computes the standard error of the values in \code{x}. If \code{na.rm} is TRUE then missing values are removed before computation proceeds.
}
\examples{
se(1:10)

}
\author{
Matthew A. Birk, \email{matthewabirk@gmail.com}
}
\seealso{
\code{\link{sd}}, \code{\link{var}}
}

